# swagger.ori.json
{
    "swagger": "2.0",
    "info": {
        "description": "Dokumentasi API Service Profil",
        "version": "1.0.0",
        "title": "Service Profil",
        "contact": {
            "name": "Racka Pratama",
            "url": "@rackapratamaps",
            "email": "racka@indonesian-aerospace.com"
        }
    },
    "host": "10.1.94.235:8002",
    "basePath": "/",
...
    "paths": {
...
        "/aplikasi/{id}/proyek": {
            "get": {
                "tags": [
                    "Aplikasi"
                ],
                "summary": "Fungsi untuk menampilkan proyek yang dimiliki aplikasi",
                "description": "Berdasarkan id aplikasi",
                "operationId": "getAllProyekUsingGET",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id aplikasi",
                        "required": true,
                        "type": "string",
                        "x-example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/Response Proyek"
                        }
                    },
                    "400": {
                        "description": "Id tidak valid"
                    },
                    "404": {
                        "description": "Proyek tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
...
        "/otoritas": {
            "get": {
                "tags": [
                    "Otoritas"
                ],
                "summary": "Fungsi untuk menampilkan otoritas",
                "operationId": "getAllUsingGET_3",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "aktif",
                        "in": "query",
                        "description": "Filter status aktif, kosongkan untuk mengambil semua otoritas, aktif = 1 jika hanya yang aktif, aktif = 0 jika semua yang nonaktif",
                        "required": false,
                        "type": "string",
                        "allowEmptyValue": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Response Otoritas"
                            }
                        }
                    },
                    "404": {
                        "description": "Otoritas tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
        "/otoritas/hapus": {
            "delete": {
                "tags": [
                    "Otoritas"
                ],
                "summary": "Fungsi untuk menghapus otoritas",
                "operationId": "deleteUsingDELETE_3",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Request Id",
                        "description": "Format JSON untuk request id",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Request Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/Response Umum"
                        }
                    },
                    "400": {
                        "description": "{param} {keterangan error}"
                    },
                    "404": {
                        "description": "Tidak ditemukan Otoritas dengan Id {id}"
                    },
                    "500": {
                        "description": "Hapus semua entitas anak dari Otoritas dengan Id {id}"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
...
        "/otoritas/tambah": {
            "post": {
                "tags": [
                    "Otoritas"
                ],
                "summary": "Fungsi untuk menambah otoritas",
                "operationId": "saveUsingPOST_3",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Request Otoritas",
                        "description": "Format JSON untuk request otoritas",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Request Otoritas"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/Response Otoritas"
                        }
                    },
                    "400": {
                        "description": "{param} {keterangan error}"
                    },
                    "409": {
                        "description": "Otoritas dengan kode {kode} sudah terdaftar"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
        "/otoritas/ubah": {
            "put": {
                "tags": [
                    "Otoritas"
                ],
                "summary": "Fungsi untuk mengubah otoritas",
                "operationId": "updateUsingPUT_3",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "Request Otoritas",
                        "description": "Format JSON untuk request otoritas",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Request Otoritas"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/Response Otoritas"
                        }
                    },
                    "400": {
                        "description": "{param} {keterangan error}"
                    },
                    "404": {
                        "description": "Tidak ditemukan Otoritas dengan Id {id}"
                    },
                    "409": {
                        "description": "Otoritas dengan kode {kode} sudah terdaftar"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
        "/otoritas/{id}": {
            "get": {
                "tags": [
                    "Otoritas"
                ],
                "summary": "Fungsi untuk menampilkan otoritas",
                "description": "Berdasarkan id otoritas",
                "operationId": "getOneUsingGET_2",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id otoritas",
                        "required": true,
                        "type": "string",
                        "x-example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "$ref": "#/definitions/Response Otoritas"
                        }
                    },
                    "400": {
                        "description": "Id tidak valid"
                    },
                    "404": {
                        "description": "Otoritas tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        },
...
        "/pengguna/{id}/proyek": {
            "get": {
                "tags": [
                    "Pengguna"
                ],
                "summary": "Fungsi untuk menampilkan proyek yang dimiliki pengguna",
                "description": "Berdasarkan id pengguna",
                "operationId": "getAllProyekUsingGET_1",
                "produces": [
                    "*/*"
                ],
                "parameters": [
                    {
                        "name": "aktif",
                        "in": "query",
                        "description": "Filter status aktif, kosongkan untuk mengambil semua proyek, aktif = 1 jika hanya yang aktif, aktif = 0 jika semua yang nonaktif",
                        "required": false,
                        "type": "string",
                        "allowEmptyValue": false
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id pengguna",
                        "required": true,
                        "type": "string",
                        "x-example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Response Proyek"
                            }
                        }
                    },
                    "400": {
                        "description": "Id tidak valid"
                    },
                    "404": {
                        "description": "Proyek tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "Bearer": [
                            "global"
                        ]
                    }
                ],
                "deprecated": false
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}